requirements.txt
langchain==0.0.336
cohere==4.34
weaviate-client==3.25.3
llama-index==0.9.1
numpy==1.22.4
pandas==1.5.2
Pillow==10.1.0
streamlit==1.28.1
streamlit-chat==0.1.1